finetune/requirements.txt
torch 
typing
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
accelerate==0.21.0 
peft==0.4.0 
bitsandbytes==0.40.2 
transformers==4.31.0 
trl==0.4.7 
sentencepiece
datasets
trl
peft
pytorch-cuda 
python-dotenv